Address remaining telemetry review comments
Widen prompt/generated token accounting to 64 bits end-to-end, document intentional Android telemetry permissions, and enforce O_NOFOLLOW for both persisted device-ID reads to close symlink TOCTOU windows.

Token width: prevents beam-weighted counters from truncating before 1DS int64 properties.
Android permissions: documents manifest-merging behavior and the telemetry-free alternative.
Device ID reads: uses bounded descriptor reads with OS-enforced no-follow for initial and concurrent-winner paths.

Copilot-Session: 6b269873-a7bf-490e-a412-70a920868135

+++ src/generators.cpp
@@ -801,7 +801,8 @@ void Generator::GenerateNextToken() {
     extra_inputs_.clear();
     transducer_state_->StepToken();
 
-    generation_telemetry_.OnTokenGenerated(state_->params_->BatchBeamSize());
+    generation_telemetry_.OnTokenGenerated(
+        static_cast<int64_t>(state_->params_->BatchBeamSize()));
     return;
   }
 
@@ -852,7 +853,8 @@ void Generator::GenerateNextToken() {
 
   last_action_ = Action::generated;
 
-  generation_telemetry_.OnTokenGenerated(search_->params_->BatchBeamSize());
+  generation_telemetry_.OnTokenGenerated(
+      static_cast<int64_t>(search_->params_->BatchBeamSize()));
 
   switch (sampling_method_) {
     case SamplingMethod::kGreedy:
